feat(keep shared things in param.rs, move others to module in which they're used)
